Moved JSON example tests
